* move selection related code to the SelectExtension   * still keep the methods as easy access within the FastAdapter --> may be revisited in the future
